backend-ml/requirements.txt
# Core ML Dependencies
tensorflow==2.15.0
numpy==1.24.3
scipy==1.11.3
matplotlib==3.7.2
pandas==2.0.3
pydicom==2.4.3
tqdm==4.66.1
scikit-learn==1.3.0

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Async Processing
celery==5.3.4
redis==5.0.1

# File Handling
Pillow==10.0.1
opencv-python-headless==4.8.1.78

# Data Validation
pydantic==2.5.0

# Logging & Monitoring
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
isort==5.12.0
mypy==1.7.1
